Move types from Map to types file
